\name{regpoy}
\alias{regpoy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Polychoric Correlation
}
\description{
Using iterative regression method to estimate the polychoric correlation between two ordinal variables or from their contingency table, under the
  assumption that the ordinal variables dissect continuous latent variables that are bivariate normal.
}
\usage{
regpoy(x, y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{the first ordered categorical variable.}
  \item{y}{a second ordered categorical variable.}
}
}

